videorender/requirements.txt
apiclient==1.0.3
appdirs==1.4.3
APScheduler==3.6.3
backcall==0.1.0
backoff==1.8.0
bcrypt==3.1.6
beautifulsoup4==4.8.2
bleach==3.1.0
bokeh==1.1.0
botocore==1.12.223
breadability==0.1.20
bs4==0.0.1
cachetools==3.0.0
cairocffi==1.1.0
CairoSVG==2.4.0
canvas==0.2
cchardet==2.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
cmudict==0.4.3
cobble==0.1.3
colorama==0.4.1
coloredlogs==14.0
colorlover==0.3.0
comtypes==1.1.7
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
cssselect2==0.2.1
cufflinks==0.15
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
Dajax==1.3
dask==1.1.3
dataclasses==0.6
dateparser==0.7.0
db==0.1.1
db-sqlite3==0.0.1
dill==0.2.9
distro==1.3.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.1.4
django-allauth==0.39.1
django-angular==2.2
django-apscheduler==0.3.0
django-chartjs==1.3
django-ckeditor==5.7.1
django-cors-headers==3.2.1
django-countries==5.5
django-crispy-forms==1.7.2
django-dajax==0.9.2
django-dajaxice==0.7
django-debug-toolbar==2.2
django-el-pagination==3.2.4
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-modelcluster==5.0.1
django-mongoengine==0.3.3
django-mysql==3.0.0.post1
django-recaptcha==2.0.4
django-rest-auth==0.9.5
django-robots==3.1.0
django-taggit==1.2.0
django-tinymce==2.8.0
django-treebeard==4.3.1
django-weasyprint==0.5.5
django-wkhtmltopdf==3.2.0
django-zipfile==0.4.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==0.4.2
djongo==1.2.33
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
docx2python==1.23
docxcompose==1.0.0a16
donut==0.2.2
draftjs-exporter==2.1.7
elasticsearch==6.3.1
en-core-web-sm==2.0.0
entrypoints==0.3
extract-social-media==0.4.0
fake-useragent==0.1.11
fb-down==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
ffmpeg-python==0.2.0
file-read-backwards==2.0.0
fpdf==1.7.2
future==0.18.2
gast==0.2.2
gensim==3.8.0
google==2.0.1
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-speech==0.36.3
google-pasta==0.1.8
google-search==1.0.2
googleapis-common-protos==1.5.6
GrabzIt==3.4.1
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.9.0
HarfPy==0.8
html-to-etree==0.2.0
html5lib==1.0.1
httpie==1.0.2
httplib2==0.12.0
humanfriendly==7.1.1
hyperlink==18.0.0
idna==2.7
image==1.5.27
imageio==2.5.0
imageio-ffmpeg==0.2.0
img2pdf==0.3.3
importlib-metadata==0.20
imutils==0.5.2
incremental==17.5.0
instagram-scraper==1.5.41
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
keyboard==0.13.3
kiwisolver==1.0.1
kwconfig==1.1.7
l18n==2018.5
lazy-object-proxy==1.3.1
lexrank==0.1.0
lxml==4.2.5
magickwand==0.2
mammoth==1.4.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
miniutils==1.0.1
mistune==0.8.4
mod-wsgi==4.6.8
MongoAlchemy==0.19
mongoengine==0.18.0
more-itertools==7.2.0
Morfessor==2.0.4
motor==2.0.0
moviepy==1.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
mysql-connector==2.1.6
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
narrative==0.1.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.2
nltk==3.4
notebook==5.7.8
numbers-extractor==0.2.3
numpy==1.17.3
oauth2client==4.1.3
oauthlib==2.1.0
opencv-python==4.0.0.21
opt-einsum==3.1.0
packaging==19.0
pandas==0.23.4
pandas-datareader==0.7.0
pandoc==1.0.2
pandocfilters==1.4.2
parse==1.11.1
parsel==1.5.1
parso==0.3.1
path.py==12.0.1
patsy==0.5.1
pdfkit==0.6.1
pep8==1.7.1
phonenumbers==8.10.6
pickleshare==0.7.5
Pillow==5.4.1
ping-cmd==1.5
pipenv==2018.11.26
pisa==3.0.33
plac==0.9.6
playsound==1.2.2
plotly==4.5.0
plotly-express==0.4.1
ply==3.11
pointofview==0.3.0
polyglot==16.7.4
pprint==0.1
pragma==0.1.0
preshed==2.0.1
proglog==0.1.9
prometheus-client==0.6.0
promise==2.2.1
prompt-toolkit==2.0.7
prosegrinder==0.1.2
protobuf==3.6.1
psutil==5.7.0
py3-validate-email==0.1.11
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
PyAutoGUI==0.9.41
pyblur==0.2.3
pycairo==1.19.1
pycld2==0.31
pycode-similar==1.2
PyContracts==1.8.12
pycountry==19.8.18
pycparser==2.19
PyDispatcher==2.0.5
pydotplus==2.0.2
pyee==5.0.0
pyfiglet==0.8.post1
PyGetWindow==0.0.4
pyglot==0.1.1
Pygments==2.3.1
PyHamcrest==1.9.0
PyICU==2.3.1
pyIsEmail==1.3.2
PyJWT==1.7.1
pylint==2.2.2
pymongo==3.8.0
PyMsgBox==1.0.6
PyMySQL==0.9.3
pyonlinedict==1.0.6
pyOpenSSL==18.0.0
pypandoc==1.4
pyparsing==2.3.0
PyPDF2==1.26.0
Pyphen==0.9.5
pypiwin32==223
pyppeteer==0.0.25
pyquery==1.4.0
pyreadline==2.1
PyRect==0.1.4
pyrsistent==0.15.4
PyScreeze==0.1.19
pyseoanalyzer==3.1.1
PySocks==1.6.8
pyTelegramBotAPI==3.6.7
pytesseract==0.2.6
python-binance==0.7.1
python-dateutil==2.7.5
python-decouple==3.3
python-docx==0.8.10
python-dotenv==0.10.3
python-telegram-bot==12.5.1
python3-openid==3.1.0
pyttsx3==2.71
pytube==9.4.0
PyTweening==1.0.3
pytz==2018.7
pyutilib.common==3.0.7
pyutilib.component.config==3.8
pyutilib.component.core==4.6.4
pyutilib.component.executables==3.5
pyutilib.misc==5.9.1
pyutilib.services==3.4
pyutilib.subprocess==3.6.2
pyvoc==1.1.0
PyWavelets==1.0.2
pywin32==224
pywinauto==0.6.5
pywinpty==0.5.5
PyXB==1.2.6
PyYAML==3.13
pyzmq==18.0.1
queuelib==1.5.0
readability==0.3.1
readability-lxml==0.7.1
regex==2018.1.10
reportlab==3.5.20
repoze.lru==0.7
requests==2.20.1
requests-file==1.4.3
requests-html==0.10.0
requests-oauthlib==1.0.0
retrying==1.3.3
rsa==4.0
Rx==1.6.1
s3transfer==0.2.1
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
Scrapy==1.5.1
seaborn==0.9.0
search-google==1.2.1
selenium==3.141.0
Send2Trash==1.5.0
seolib==0.1.3
service-identity==17.0.0
signals==0.0.2
simple-geoip==0.1.0
simplejson==3.16.0
singledispatch==3.4.0.3
siterank==1.0.3
six==1.11.0
sklearn==0.0
smart-open==1.8.4
sounddevice==0.3.12
soupsieve==2.0
spacy==2.0.18
SpeechRecognition==3.8.1
SQLAlchemy==1.3.4
sqlparse==0.2.4
statsmodels==0.11.0
stripe==2.43.0
summa==1.2.0
sumy==0.8.1
svgwrite==1.2.1
syllables==0.1.0
tabula-py==1.3.1
telnetlib3==1.0.2
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-summarizer==0.0.4
textblob==0.15.2
textblob-fr==0.2.0
textstat==0.5.6
textwrap3==0.9.1
Theano==1.0.4
thinc==6.12.1
timeout-decorator==0.4.1
tinycss2==1.0.2
tld==0.9.2
tldextract==2.2.2
tlds==2019022700
toolz==0.9.0
tornado==6.0.2
tqdm==4.28.1
traitlets==4.3.2
tweepy==3.7.0
Twisted==18.9.0
txaio==18.8.1
type-spec==0.3
tzlocal==1.5.1
ua-parser==0.10.0
uharfbuzz==0.4.0
ujson==1.35
unicodedata2==11.0.0
Unidecode==1.1.0
uniseg==0.7.1
uritemplate==3.0.0
uritools==2.2.0
urlextract==0.13.0
urllib3==1.24.1
urlpath==1.1.4
user-agents==2.1
validate-email==1.3
verify-email==2.3.1
virtualenv==16.1.0
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
w3lib==1.19.0
wagtail==2.9
wagtailfontawesome==1.2.1
wcwidth==0.1.7
WeasyPrint==51
webdriverdownloader==0.1.dev2
webencodings==0.5.1
webguitest-scubbx==0.0.15
websocket-client==0.56.0
websockets==7.0
Werkzeug==0.15.4
wfastcgi==3.0.0
wget==3.2
widgetsnbextension==3.4.2
Willow==1.3
windows-curses==2.0
wit==5.1.0
wordcloud==1.5.0
wpdetect==1.3.3
wrapt==1.11.2
WTForms==2.2.1
xhtml2pdf==0.2.3
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
yapf==0.27.0
youtube==1.0.6
youtube-data-api==0.0.15
youtube-python==1.0.12
youtube-transcript-api==0.1.2
zipp==0.6.0